backend/requirements.txt
# TradeTracker Backend
# FastAPI + PostgreSQL

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Data Processing (from main project)
pandas>=2.1.0
pydantic>=2.5.0
python-dateutil>=2.8.2

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# CORS & Security
python-dotenv>=1.0.0

# Async utilities
aiofiles>=23.2.1

